requirements.txt
streamlit
openai
python-dotenv
boto3
openai-agents
asyncio
fastapi
uvicorn
requests
python-multipart
pymongo
astrapy